Create Menger_4D
Build a 4D Menger sponge via recursive integer subdivision.
At each level, the hypercube is subdivided into 3^4 sub-hypercubes. For 4D, we remove those sub-hypercubes in which three or more of the indices equal 1. (i.e. we keep a sub-hypercube if the count of ones in (i,j,k,l) is <= 2.)